+++ jobbergate-api/jobbergate_api/apps/job_scripts/tools.py
@@ -1,5 +1,7 @@
 """Provide a convenience class for managing job-script files."""
 
+import re
+
 from loguru import logger
 
 
@@ -15,16 +17,22 @@ def inject_sbatch_params(job_script_data_as_string: str, sbatch_params: list[str
         logger.warning("Sbatch param list is empty")
         return job_script_data_as_string
 
-    first_sbatch_index = job_script_data_as_string.find("#SBATCH")
-    string_slice = job_script_data_as_string[first_sbatch_index:]
-    line_end = string_slice.find("\n") + first_sbatch_index + 1
+    # Find the first non-blank, non-comment line
+    match = re.search(r"^[^#\n]", job_script_data_as_string, re.MULTILINE)
+    if match:
+        insert_index = match.start()
+    else:
+        # If no such line is found, append at the end
+        insert_index = len(job_script_data_as_string)
 
-    inner_string = ""
+    inner_string = "# Sbatch params injected at rendering time\n"
     for parameter in sbatch_params:
-        inner_string += "#SBATCH " + parameter + "\n"
+        inner_string += f"#SBATCH {parameter}\n"
+    else:
+        inner_string += "\n"
 
     new_job_script_data_as_string = (
-        job_script_data_as_string[:line_end] + inner_string + job_script_data_as_string[line_end:]
+        job_script_data_as_string[:insert_index] + inner_string + job_script_data_as_string[insert_index:]
     )
 
     logger.debug("Done injecting sbatch params into job script")